refactor: reorganize imports and remove unnecessary awaits for getCurrentTier
