+++ src/app/api/chat/public/route.ts
@@ -5,8 +5,8 @@ import { NextResponse } from 'next/server';
 import * as fs from 'fs/promises';
 import * as path from 'path';
 import { getWalletByUserId } from '@/lib/wallet';
-import { getUserRank } from '@/lib/ranks';
-import { getBotTierSettings, getCurrentTier } from '@/lib/tiers';
+import { getUserRank, getCurrentTier } from '@/lib/ranks';
+import { getBotTierSettings } from '@/lib/tiers';
 import type { Rank } from '@/lib/ranks';
 import type { TierSetting } from '@/lib/tiers';
 
@@ -68,7 +68,7 @@ export async function POST(request: Request) {
         }
 
         const rank = getUserRank(wallet.balances.usdt || 0);
-        const tier = await getCurrentTier(wallet.balances.usdt || 0, tierSettings);
+        const tier = getCurrentTier(wallet.balances.usdt || 0, tierSettings);
 
         const newMessage: ChatMessage = {
             id: `msg_${Date.now()}_${Math.random()}`,
