+++ src/app/api/squad/route.ts
@@ -1,7 +1,6 @@
 
 import { NextResponse } from 'next/server';
-import { getUserRank } from '@/lib/ranks';
-import { getCurrentTier } from '@/lib/tiers';
+import { getUserRank, getCurrentTier } from '@/lib/ranks';
 import { getBotTierSettings } from '@/lib/tiers';
 
 const MOCK_USER_ID = 'mock-user-123';
